pyproject.toml
[tool.poetry]
name = "online-school"
version = "0.1.0"
description = ""
authors = ["Katsiarynna <kotboruna@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
channels = "^4.0.0"
channels-redis = "^4.2.0"
mentorship = "^0.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
